requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.89.0
anyio==4.13.0
APScheduler==3.10.4
attrs==26.1.0
Authlib==1.6.9
banks==2.4.1
beautifulsoup4==4.14.3
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
colorama==0.4.6
cryptography==46.0.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
ecdsa==0.19.2
fastapi==0.135.1
fastuuid==0.14.0
filelock==3.25.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.3.0
geographiclib==2.1
geopy==2.4.1
google-adk @ git+https://github.com/google/adk-python.git@1445ad5069841e446328e0856553f69a6699f0f4
google-ai-generativelanguage==0.6.10
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-cloud-aiplatform==1.145.0
google-cloud-appengine-logging==1.9.0
google-cloud-audit-log==0.5.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.26.0
google-cloud-bigtable==2.36.0
google-cloud-core==2.5.1
google-cloud-dataplex==2.18.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.22.0
google-cloud-logging==3.15.0
google-cloud-monitoring==2.30.0
google-cloud-pubsub==2.23.0
google-cloud-resource-manager==1.17.0
google-cloud-secret-manager==2.27.0
google-cloud-spanner==3.64.0
google-cloud-speech==2.38.0
google-cloud-storage==2.18.2
google-cloud-trace==1.19.0
google-crc32c==1.8.0
google-genai==1.70.0
google-generativeai==0.8.3
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
graphviz==0.21
greenlet==3.3.2
griffe==2.0.2
griffecli==2.0.2
griffelib==2.0.2
grpc-google-iam-v1==0.14.4
grpc-interceptor==0.15.4
grpcio==1.80.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.9.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
k8s-agent-sandbox==0.2.1
kubernetes==35.0.0
langchain-core==1.2.26
langgraph==0.4.7
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.26
litellm==1.82.6
llama-index-core==0.14.20
llama-index-embeddings-google-genai==0.5.0
llama-index-instrumentation==0.5.0
llama-index-readers-file==0.6.0
llama-index-workflows==2.17.1
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.27.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.4
oauthlib==3.3.1
openai==2.30.0
opentelemetry-api==1.38.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pillow==10.4.0
platformdirs==4.9.4
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pyOpenSSL==26.0.0
pyparsing==3.3.2
pypdf==6.9.2
PyPika==0.51.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.24
pytz==2026.1.post1
# pywin32==311  <-- Windows-only, removed for Linux/Cloud Run compatibility
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.30.0
rsa==4.9.1
setuptools==82.0.1
shapely==2.0.5
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
sqlalchemy-spanner==1.17.3
sqlparse==0.5.5
sse-starlette==3.3.4
starlette==0.52.1
staticmap==0.5.7
striprtf==0.0.26
tenacity==9.1.4
tiktoken==0.12.0
tinytag==2.2.1
tokenizers==0.22.2
toolbox-adk==1.0.0
toolbox-core==1.0.0
tqdm==4.67.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0